+++ CHANGELOG.md
@@ -16,7 +16,7 @@ As this project is a user-facing application, the places in the semantic version
 ### Added
 
 - It is now possible to add multiple terms to the query, and they are combined with AND
-- Show Queerlit motivation text in search result
+- Show Queerlit motivation text and peripheral terms in search result
 
 ### Changed